Expand `TurboNavigationDelegate` to support loading Session's Web View
Access the `WKWebView` instance when it gets loaded by the `Session`.

Being able to access the `WebView` when it's loaded might be useful to
your application. For example, you can attach a

```swift
class MyCustomClass: TurboNavigationDelegate, NSObject, WKScriptMessageHandler {
    let navigator = TurboNavigator(delegate: self)

    func session(_ session: Session, didLoadWebView webView: WKWebView) {
        webView.configuration.userContentController.add(self, name: "myNativeApp")
    }

    func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage ) {
        // handle the JavaScript Object message
    }
}
```

With the `session` and `modalSession` properties marked with `internal`
access, application's are forced to construct their own `WKWebView`
instances by setting `TurboConfig.shared.makeCustomWebView`.

By expanding the `TurboNavigationDelegate` protocol to notify delegates
when the `WKWebView` is loaded, it exposes _both_ the `Session` instance
_and_ the `WKWebView` instance.