Support other casings for `view-controller`
The decision to use dash-casing for the `view-controller` property
matches the multi-word precedent set by the [Turbo iOS examples][].
However, the [Turbo Android examples][] uses snake_casing.

It's unlikely that an Android app and an iOS app will share exact copies
of a path configuration file's structure and data. However, it's
possible that they'd share a similar server-side backend to generate
those files, or a shared naming convention across those files.

The `view-controller` appears to be the only multi-word key expected and
supported out-of-the-box, with the expectation that other keys can be
used to inform bespoke features in app-specific contexts.

This commit expands that support to match other naming conventions,
mainly snake_case and camelCase.

[Turbo iOS examples]: https://github.com/hotwired/turbo-ios/blob/c9d060fcdec136c3808cd0764e956d375012c6fb/Docs/PathConfiguration.md#path-configuration
[Turbo Android examples]: https://github.com/hotwired/turbo-android/blob/bf945e18aed3d5749c9efc6e228d98bc012878af/docs/PATH-CONFIGURATION.md#path-configuration